requirements.txt
bleach==1.5.0
cffi==1.12.3
cycler==0.10.0
enum34==1.1.6
html5lib==0.9999999
kiwisolver==1.0.1
Markdown==3.1
matplotlib==3.0.3
numpy==1.16.3
opencv-python==4.1.0.25
Pillow==6.0.0
protobuf==3.7.1
pycparser==2.19
pygame==1.9.6
pymunk==5.4.2
pyparsing==2.4.0
PyQt5==5.12.1
PyQt5-sip==4.19.15
python-dateutil==2.8.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
sklearn==0.0
tensorflow-gpu
tensorflow-tensorboard
Werkzeug==0.15.2

# TODO: update!